Bug 547864 - [SequenceDiagram] Undo/Redo problem with CombinedFragment
Failing tests :
  - Some tests are failing due to incorrect tests (validation tests are
no more needed because the covering is automatically recalculated)
  - Other tests are failing due to undo/redo problem of CombinedFragment
creation. The undo creation when something is contained into the created
CombinedFragment crash with exception (The undo of creation is no more
possible because the elements positions changed).

In this patch:
  I manage undo/redo of creation differently (manually) to move needed
contained elements before to remove CombinedFragment. In this case, i
have to cut the undo/redo of the SemanticCreationCommand.

With this patch, all JUnit tests of sequence diagram works fine.

Change-Id: Ib1ed2df10404763ca79e5a2c5c62bf4c52600657
Signed-off-by: Nicolas FAUVERGUE <nicolas.fauvergue@cea.fr>